eval: Move some decoding functions to eval/decode.c
